moved projectRequirements state into projectReducer
