+++ app/assets/javascripts/redux/reducers/projectReducer.jsx
@@ -1,18 +1,28 @@
 const projectReducer = (state = {
     projects: [],
-    project: []
+    project: [],
+    projectRequirements: []
 }, action) => {
     switch (action.type) {
         case "GET_ALL_PROJECTS":
             state = {
                 projects: action.payload,
-                project: state.project
+                project: state.project,
+                projectRequirements: state.projectRequirements
             };
             break;
         case "GET_PROJECT_BY_ID":
             state = {
                 projects: state.projects,
-                project: action.payload
+                project: action.payload,
+                projectRequirements: state.projectRequirements
+            };
+            break;
+        case "GET_REQUIREMENTS_BY_PROJECT_ID":
+            state = {
+                projectRequirements: action.payload,
+                projects: state.projects,
+                project: state.project
             };
             break;
     }
